♻️ refactor: 라우터 컴포넌트로 분리
